+++ src/squirrel/squirrel_util.cpp
@@ -155,13 +155,13 @@ void print_squirrel_stack(HSQUIRRELVM v)
       case OT_INTEGER: {
         SQInteger val;
         sq_getinteger(v, i, &val);
-        printf("integer (%d)", static_cast<int> (val));
+        printf("integer (%d)", static_cast<int>(val));
         break;
       }
       case OT_FLOAT: {
         SQFloat val;
         sq_getfloat(v, i, &val);
-        printf("float (%f)", val);
+        printf("float (%f)", static_cast<double>(val));
         break;
       }
       case OT_STRING: {
